LPS-57255 Move beans
