+++ portal-service/src/com/liferay/portlet/dynamicdatamapping/DDMForm.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.dynamicdatamapping.model;
+package com.liferay.portlet.dynamicdatamapping;
 
 import java.io.Serializable;
