--- portal-service/src/com/liferay/portlet/dynamicdatamapping/model/packageinfo
@@ -1,1 +1,0 @@
-version 7.0.0
